refactor: :bulb: organize components
